backend/requirements.txt
Flask>=2.0
flask-cors>=3.0
pytest>=7.0
Pillow>=10.0
pymongo==3.12
pymongo[srv]==3.12
python-dotenv>=1.0
google-genai==1.39.1
python-jose==3.3.0
requests>=2.31